+++ R/ftExtra.R
@@ -1,3 +1,4 @@
+#' @importFrom rlang .data
 #' @importFrom magrittr %>%
 #' @export
 magrittr::`%>%`